// src/types/AccountResponse.tsx
export interface AccountResponse {
  employeeId: number;
  employeeCode: string;
  firstName: string;
  lastName: string;
  gender: string;
  birthDate: string;
  hireDate: string;
  email: string;
  phoneNumber: string;
  departmentName: string;
  positionName: string;
  salary: number;
  status: string;
  userId: number;
  username: string;
  role: string;
  profilePictureUrl: string;
  presignedRequestUrl: string;
}